Add gallery component + gallery api endpoint
This commit adds a fs parse into the server.js which parses the gallery
image directory and returns an array in json format to the api endpoint.
The gallery component in turn can use this json formatted array to
display the images sitting inside the static gallery directory to the
client.